+++ guillotina/behaviors/dublincore.py
@@ -95,10 +95,6 @@ class DublinCore(AnnotationBehavior):
     creation_date = ContextProperty('creation_date', None)
     modification_date = ContextProperty('modification_date', None)
 
-    # all properties but these 4 are not annotated
-    __local__properties__ = ('creation_date', 'modification_date',
-                             'creators', 'contributors', 'title')
-
     def __init__(self, context):
         self.__dict__['context'] = context
         super(DublinCore, self).__init__(context)