+++ guillotina/behaviors/properties.py
@@ -1,8 +1,10 @@
 
+_EMPTY = object()
+
 
 class ContextProperty:
 
-    def __init__(self, attribute, default):
+    def __init__(self, attribute, default=_EMPTY):
         self.__name__ = attribute
         self.default = default
 
@@ -13,6 +15,11 @@ class ContextProperty:
         result = getattr(inst.context, self.__name__, self.default)
         if callable(result):
             result = result(context=inst.context, name=self.__name__)
+        if result == _EMPTY:
+            try:
+                return inst.schema[self.__name__].default
+            except KeyError:
+                raise AttributeError(self.__name__)
         return result
 
     def __set__(self, inst, value):